+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.0] - 2020-03-25
+
 ### Added
 - Export interfaces of cnd lightning body requests.
 
@@ -180,7 +182,8 @@ The SDK now recognizes this: Promises returned from calls on `Cnd` will now reje
 ### Added
 - Common code that can be used to build applications on top of COMIT.
 
-[Unreleased]: https://github.com/comit-network/comit-js-sdk/compare/0.14.1...HEAD
+[Unreleased]: https://github.com/comit-network/comit-js-sdk/compare/0.15.0...HEAD
+[0.15.0]: https://github.com/comit-network/comit-js-sdk/compare/0.14.1...0.15.0
 [0.14.1]: https://github.com/comit-network/comit-js-sdk/compare/0.14.0...0.14.1
 [0.14.0]: https://github.com/comit-network/comit-js-sdk/compare/0.13.0...0.14.0
 [0.13.0]: https://github.com/comit-network/comit-js-sdk/compare/0.12.0...0.13.0